# API/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Ecommerce API",
        "description": "Tài liệu API cho hệ thống Ecommerce",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/v1/customers/orders/previews": {
            "post": {
                "tags": [
                    "Customer / Order"
                ],
                "summary": "Preview checkout totals, shipping & discounts",
                "operationId": "5038a8a31ec9d3b2dac4919fc27eef3c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CheckoutRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preview data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Preview checkout successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/customers/orders": {
            "post": {
                "tags": [
                    "Customer / Orders"
                ],
                "summary": "Place the order",
                "operationId": "60e4867225186b3fe6dd6e3aa458d866",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlaceOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order placed successfully."
                                        },
                                        "data": {}
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/customers/orders/{id}": {
            "get": {
                "tags": [
                    "Customer / Orders"
                ],
                "summary": "Get single order details",
                "operationId": "b84d6295af66e155a28efa276f48caea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SubOrderResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CheckoutRequest": {
                "required": [
                    "shipping_address_id",
                    "billing_address_id",
                    "shipping_method",
                    "payment_method"
                ],
                "properties": {
                    "shipping_address_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "billing_address_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "shipping_method": {
                        "type": "string",
                        "example": "standard"
                    },
                    "payment_method": {
                        "type": "string",
                        "example": "cod"
                    },
                    "coupon_code": {
                        "type": "string",
                        "example": "DISCOUNT10",
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "example": "Please ring the bell",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PlaceOrderRequest": {
                "required": [
                    "shipping_address_id",
                    "billing_address_id",
                    "shipping_method",
                    "payment_method"
                ],
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/CheckoutRequest"
                    },
                    {
                        "properties": {
                            "payment_method": {
                                "type": "string",
                                "example": "cod"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "SubOrderResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 123
                    },
                    "order_number": {
                        "type": "string",
                        "example": "ORD-20250724-000123"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 42
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 499000
                    },
                    "shipping_fee": {
                        "type": "number",
                        "format": "float",
                        "example": 15000
                    },
                    "discount": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 464000
                    },
                    "payment_method": {
                        "type": "string",
                        "example": "cod"
                    },
                    "payment_status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "payment_reference": {
                        "type": "string",
                        "example": "TXN-XYZ789",
                        "nullable": true
                    },
                    "shipping_method": {
                        "type": "string",
                        "example": "standard"
                    },
                    "shipping_address": {
                        "type": "object"
                    },
                    "billing_address": {
                        "type": "object"
                    },
                    "status": {
                        "type": "string",
                        "example": "processing"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Gift wrap please",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date",
                        "example": "24-07-2025"
                    },
                    "sub_orders": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SubOrderResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}